+++ CHANGELOG.md
@@ -36,6 +36,10 @@
 - in ``repl.arsh`` module, load history at first prompt
 - add ``PRE_EXECS`` hook for call functions before executing command
 
+### Fixed
+
+- bugfix method lookup for nested named imported types
+
 ## [0.39.3] - 2025-10-31
 
 - fix nested sub-command lookup